rename dhis2 authenticator module for consistency
